split util.py into util/{graphs,random_variables,tensorflow}.py
