+++ edward/util/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from edward.util.graphs import *
+from edward.util.random_variables import *
+from edward.util.tensorflow import *
